Estimates the distributions using mle.

estimate_pars estimates the distribution's parameters using their maximum likelihood estimator. This estimate is conditional on the current xmin value.

Usage

estimate_pars(m, pars = NULL)

Arguments

m

A reference class object that contains the data.

pars

default NULL. A vector or matrix (number of columns equal to the number of parameters) of parameters used to #' optimise over. Otherwise, for each value of xmin, the mle will be used, i.e. estimate_pars(m). For small samples, the mle may be biased.

Value

returns list.

Examples

data(moby_sample)
m = displ$new(moby_sample)
estimate_xmin(m)
#> $gof
#> [1] 0.01554137
#> 
#> $xmin
#> [1] 7
#> 
#> $pars
#> [1] 1.957412
#> 
#> $ntail
#> [1] 335
#> 
#> $distance
#> [1] "ks"
#> 
#> attr(,"class")
#> [1] "estimate_xmin"
m$setXmin(7)
estimate_pars(m)
#> $pars
#> [1] 1.957411
#> 
#> $value
#> [1] 1327.789
#> 
#> $counts
#> function gradient 
#>        5        5 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> [1] "CONVERGENCE: REL_REDUCTION_OF_F <= FACTR*EPSMCH"
#> 
#> attr(,"class")
#> [1] "estimate_pars"