Values greater than or equal to xmin

Returns the number of data points greater than or equal to current value of xmin. In the Clauset et al, paper this is called ntail.

Usage

get_ntail(m, prop = FALSE, lower = FALSE)

Arguments

m

a distribution object.

prop

default FALSE. Return the value as a proportion of the total sample size

lower

default FALSE. If TRUE returns sample size - ntail

Examples

################################################
# Load data and create example object
################################################
data(moby_sample)
m = displ$new(moby_sample)
m$setXmin(7)
################################################
# Get ntail
################################################
get_ntail(m)
#> [1] 335
sum(moby_sample >= 7)
#> [1] 335