Generic plotting functions

These are generic functions for distribution reference objects. Standard plotting functions, i.e. plot, points, and lines work with all distribution objects.

Usage

# S4 method for class 'distribution'
lines(x, cut = FALSE, draw = TRUE, length.out = 100, ...)

# S4 method for class 'distribution'
plot(x, cut = FALSE, draw = TRUE, ...)

# S4 method for class 'distribution'
points(x, cut = FALSE, draw = TRUE, length.out = 100, ...)

Arguments

x

a distribution reference object.

cut

logical (default FALSE) - Where should the plot begin. If cut=FALSE, then the plot will start at the minimum data value. Otherwise, the plot will start from xmin

draw

logical (default TRUE). Should the plot/lines/points function plot or return the data (in a data frame object).

length.out

numeric, default 100. How many points should the distribution be evaluated at. This argument is only for plotting the fitted lines.

...

Further arguments passed to the lines functions.

Note

This method does not alter the internal state of the distribution objects.